scripts/requirements.txt
python-slugify[unidecode]==6.1.1
requests==2.27.1
pandas==1.4.1
frictionless==4.28.1
